+++ persistent/Database/Persist/Types.hs
@@ -69,7 +69,6 @@ import Database.Persist.Types.Base
        , PersistValue(..)
        , ReferenceDef(..)
        , SqlType(..)
-       , Span(..)
        , UniqueDef(..)
        , UpdateException(..)
        , WhyNullable(..)